style: move SidebarMain, move SidebarStyles
