--- wondrous-app/components/Common/Sidebar/index.tsx
@@ -1,1 +1,0 @@
-export { default as SideBarComponent } from './Main';
